Move SDK related items to the ServiceInsight.Sdk project
