+++ src/ServiceInsight.Sdk/ExtensionMethods/ReflectionHelper.cs
