+++ src/ServiceInsight.Sdk/MessageViewers/ICustomMessageBodyViewer.cs
@@ -1,0 +1,10 @@
+﻿using ServiceInsight.Models;
+using ServiceInsight.ServiceControl;
+
+namespace ServiceInsight.MessageViewers
+{
+    public interface ICustomMessageBodyViewer : IDisplayMessageBody
+    {
+        bool IsVisible(StoredMessage selectedMessage, PresentationHint presentationHint);
+    }
+}
