+++ src/ServiceInsight.Sdk/MessageViewers/IDisplayMessageBody.cs
