+++ src/ServiceInsight.Sdk/Models/Endpoint.cs
