+++ src/ServiceInsight.Sdk/Models/ExceptionDetails.cs
