+++ src/ServiceInsight.Sdk/Models/FailureDetails.cs
