+++ src/ServiceInsight.Sdk/Models/HeaderInfo.cs
