+++ src/ServiceInsight.Sdk/Models/Message.cs
