+++ src/ServiceInsight.Sdk/Models/MessageInfo.cs
