+++ src/ServiceInsight.Sdk/Models/MessageIntent.cs
