+++ src/ServiceInsight.Sdk/ServiceControl/PresentationHint.cs
@@ -1,0 +1,8 @@
+﻿namespace ServiceInsight.ServiceControl
+{
+    public enum PresentationHint
+    {
+        Standard,
+        NoContent
+    }
+}
