+++ src/ServiceInsight.sln
@@ -18,6 +18,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceInsight.Licensing", "ServiceInsight.Licensing\ServiceInsight.Licensing.csproj", "{5162668F-2FB4-453B-9D2F-88D68A9286E7}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceInsight.Sdk", "ServiceInsight.Sdk\ServiceInsight.Sdk.csproj", "{7EF919F1-B497-4478-8EB4-ECDB241F4A88}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,6 +42,10 @@ Global
 		{5162668F-2FB4-453B-9D2F-88D68A9286E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5162668F-2FB4-453B-9D2F-88D68A9286E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5162668F-2FB4-453B-9D2F-88D68A9286E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7EF919F1-B497-4478-8EB4-ECDB241F4A88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7EF919F1-B497-4478-8EB4-ECDB241F4A88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7EF919F1-B497-4478-8EB4-ECDB241F4A88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7EF919F1-B497-4478-8EB4-ECDB241F4A88}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
